Move Adaptive Cards to bundle
