+++ packages/bundle/src/adaptiveCards/Attachments/AdaptiveCardAttachment.js
