+++ packages/bundle/src/adaptiveCards/Attachments/AudioCardAttachment.js
