+++ packages/bundle/src/adaptiveCards/Attachments/CommonCard.js
