+++ packages/bundle/src/adaptiveCards/Attachments/SignInCardAttachment.js
