+++ packages/bundle/src/adaptiveCards/Attachments/VideoCardAttachment.js
