+++ packages/component/package.json
@@ -58,7 +58,6 @@
     "typescript": "^3.1.6"
   },
   "dependencies": {
-    "adaptivecards": "^1.0.0",
     "botframework-webchat-core": "^0.0.0-0",
     "bytes": "^3.0.0",
     "classnames": "^2.2.6",
